Rename neet.interfaces to neet.abc
