+++ test/test_reca.py
@@ -1,6 +1,7 @@
 import unittest
 import numpy as np
 from neet.automata.reca import RewiredECA
+from neet.abc import Network
 
 
 class TestRewiredECA(unittest.TestCase):
@@ -12,7 +13,6 @@ class TestRewiredECA(unittest.TestCase):
         """
         Ensure that RewiredECA meets the requirement of a network
         """
-        from neet.interfaces import Network
         self.assertTrue(isinstance(RewiredECA(23, size=3), Network))
         self.assertTrue(isinstance(RewiredECA(
             30, wiring=[[-1, 0, 1], [0, 1, 2], [1, 2, 3]]), Network))
